More code optim
