+++ firmware/application/external/epirb_rx/location.cpp
@@ -20,40 +20,10 @@
  */
 
 #include "location.hpp"
-#include "beacon.hpp"
 #include <cstdio>
 #include "string_format.hpp"
 
 namespace ui::external_app::epirb_rx {
-
-size_t Beacon::formatSummary(char* buffer, bool with_time) {
-    size_t result = 0;
-    if (with_time) {
-        result += formatTime(buffer);
-        buffer[result++] = '-';
-    }
-    result += sprintf((buffer + result), "%4s-%5s-", shortId().c_str(), getType());
-    if (location.isUnknown()) {
-        result += sprintf((buffer + result), "      ");
-    } else {
-        result += location.toString((buffer + result), Location::LocationFormat::MAIDENHEAD_LOCATOR);
-    }
-    result += sprintf((buffer + result), "[%s%s%s]", isFrameValid() ? (bch1Corrected || bch2Corrected) ? STR_COLOR_YELLOW : STR_COLOR_GREEN : STR_COLOR_RED, getSatus().c_str(), STR_COLOR_WHITE);
-    return result;
-}
-
-bool Beacon::isFrameValid() {
-    return isBch1Valid() && ((!hasBch2) || isBch2Valid()) && (!isEmpty);
-}
-
-std::string Beacon::shortId() {
-    std::string result = std::string(hexId);
-    return (result.size() >= 4) ? result.substr(0, 4) : result;
-}
-
-size_t Beacon::formatTime(char* buffer) {
-    return sprintf(buffer, "%02d:%02d:%02d", date.hour(), date.minute(), date.second());
-}
 
 void Location::Angle::clear() {
     degrees = 255;
