+++ firmware/application/external/external.cmake
@@ -265,6 +265,7 @@ set(EXTCPPSRC
 	external/epirb_rx/ui_epirb_rx.cpp
 	external/epirb_rx/ui_beaconlist.cpp
 	external/epirb_rx/beacon_db.cpp
+	external/epirb_rx/beacon.cpp
 	external/epirb_rx/location.cpp
 
 	#epirb_tx
